refactor: consolidate duplicate telemetry functions
Move serializeTelemetryData and spawnTelemetryWorker functions from
worker-client.ts and sentry.ts to telemetry-utils.ts to eliminate
duplication. Keep isUserError in sentry.ts since it's only used there
for Sentry-specific error filtering.
