+++ src/hooks/finally/sentry.ts
@@ -1,6 +1,3 @@
-import {spawn} from 'node:child_process'
-import {fileURLToPath} from 'node:url'
-import {dirname, join} from 'node:path'
 import {Hook} from '@oclif/core/hooks'
 
 /**
@@ -32,56 +29,6 @@ function isUserError(error: any): boolean {
   return false
 }
 
-/**
- * Serialize data for telemetry worker, handling Error objects specially
- */
-function serializeTelemetryData(data: any): string {
-  // If it's an Error object, convert to plain object with all properties
-  if (data instanceof Error) {
-    return JSON.stringify({
-      // Include any other enumerable properties first
-      ...data,
-      // Then override with important properties to ensure they're captured
-      message: data.message,
-      name: data.name,
-      stack: data.stack,
-      code: (data as any).code,
-      statusCode: (data as any).statusCode,
-      http: (data as any).http,
-      oclif: (data as any).oclif,
-    })
-  }
-
-  return JSON.stringify(data)
-}
-
-/**
- * Spawn telemetry worker process in background
- * This avoids blocking the main CLI process with telemetry overhead
- */
-function spawnTelemetryWorker(data: any) {
-  try {
-    const __dirname = dirname(fileURLToPath(import.meta.url))
-    const workerPath = join(__dirname, '..', '..', '..', 'dist', 'lib', 'analytics-telemetry', 'telemetry-worker.js')
-    const child = spawn(process.execPath, [workerPath], {
-      detached: true,
-      // Keep stderr attached to see DEBUG output, but ignore stdout
-      stdio: ['pipe', 'ignore', 'inherit'],
-      // On Windows, prevent console window from appearing
-      windowsHide: true,
-    })
-
-    // Send data via stdin
-    child.stdin.write(serializeTelemetryData(data))
-    child.stdin.end()
-
-    // Detach from parent so it can exit immediately
-    child.unref()
-  } catch {
-    // Silently fail - don't let telemetry errors affect user experience
-  }
-}
-
 const finallyHook: Hook<'finally'> = async function (options) {
   // Only process if there was an error
   if (!options.error) {
@@ -93,7 +40,7 @@ const finallyHook: Hook<'finally'> = async function (options) {
     return
   }
 
-  const {isTelemetryEnabled} = await import('../../lib/analytics-telemetry/telemetry-utils.js')
+  const {isTelemetryEnabled, spawnTelemetryWorker} = await import('../../lib/analytics-telemetry/telemetry-utils.js')
 
   // Use the consolidated telemetry check
   if (!isTelemetryEnabled()) {
